force candy-machine to work
candy-machine is unable to use modules from the `common` package. i wasted
enough hours on it. this approach (copying the code in) is embarrassingly
terrible but solving the problem this ugly way is better than waiting
yet another day to ship this. will figure out what is wrong with lerna
in a follow up PR.